Files and folders renamed
Files and folders were renamed to suit its implementation in the code base
