+++ Multi Step Form/src/App.tsx
@@ -1,6 +1,6 @@
 import "./App.css";
 import { Route, Routes } from "react-router-dom";
-import FrontPage from "./components/Landing/Landing.component.tsx";
+import FrontPage from "./components/Landing/Landing.tsx";
 import AuthForm from "./components/AuthForm/AuthForm.component.tsx";
 
 function App() {
