--- Multi Step Form/src/utils/__mocks__/form.field.validation.ts
@@ -1,3 +1,0 @@
-const formFieldValidation = jest.fn();
-
-export default formFieldValidation;
