--- Multi Step Form/src/utils/__mocks__/useMultiForm.ts
@@ -1,6 +1,0 @@
-export const prevStep = jest.fn();
-export const nextStep = jest.fn();
-export const goToStep = jest.fn();
-export const isFirstStep = jest.fn();
-export const isLastStep = jest.fn();
-export const reset = jest.fn();
