+++ Multi Step Form/src/utils/redux/formSlice.ts
@@ -1,0 +1,33 @@
+// import { createSlice, PayloadAction } from "@reduxjs/toolkit";
+
+// export interface FormState {
+//   formData: FormItems;
+//   errors: {
+//     nameErr: string;
+//     emailErr: string;
+//     phoneNumberErr: string;
+//     addonsErr: string;
+//   };
+//   currentStep: number;
+//   isComplete: boolean;
+//   summaryErr: string;
+// }
+// const initialState: FormState = {
+//   formData: {
+//     name: "",
+//     email: "",
+//     phoneNumber: "",
+//     timeFrame: "monthly",
+//     plan: { price: 9, title: "arcade" },
+//     addons: [],
+//   },
+//   errors: {
+//     nameErr: "",
+//     emailErr: "",
+//     phoneNumberErr: "",
+//     addonsErr: "",
+//   },
+//   currentStep: 0,
+//   isComplete: false,
+//   summaryErr: "",
+// };
