+++ Multi Step Form/src/utils/redux/store.ts
@@ -1,0 +1,5 @@
+import { configureStore } from "@reduxjs/toolkit";
+
+export const store = configureStore({
+  reducer: {},
+});
